varstring: Prevent suffix match on key
FindKeyAssignInString uses a naive matching check that behaves unexpectedly in
the case where the match is the suffix of a different key.

Given the following varstring:

    "buff_duration=5 duration=2 max_duration=10"

Attempting to find "duration" would return the index after "buff_duration=".

The change adds a check to ensure that the matching substring is either at the
start of the varstring, or that it is preceded by the delimiter (' ').  If
neither applies, it iterates over the rest of the string.

There shouldn't be a notable performance impact in most cases since the first
match is usually correct.